MOE2/requirements.txt
altgraph==0.15
APScheduler==3.5.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==1.5.0
boto3==1.4.4
botocore==1.5.95
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
constantly==15.1.0
cx-Freeze==5.1.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
Flask==0.12.2
freetype-py==1.1
ftfy==4.4.3
future==0.16.0
gevent==1.2.1
greenlet==0.4.12
h5py==2.7.1
hexdump==3.3
html5lib==0.9999999
hyperlink==17.3.1
incremental==17.5.0
itchat==1.2.32
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.5
jsonschema==2.6.0
Keras==2.1.2
Kivy==1.10.0
Kivy-Garden==0.1.4
klein==17.2.0
lxml==4.1.1
macholib==1.9
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==1.5.3
MetOffer==1.3.2
murmurhash==0.26.4
numpy==1.13.1
oauthlib==2.0.6
pathlib==1.0.1
pefile==2017.11.5
Pillow==5.0.0
pipreqs==0.4.9
plac==0.9.6
preshed==1.0.0
protobuf==3.5.1
PyAudio==0.2.11
pydub==0.20.0
Pygments==2.2.0
PyInstaller==3.3.1
pykwalify==1.6.0
pyparsing==2.2.0
pypiwin32==222
pypng==0.0.18
PyQRCode==1.2.1
PyQt5==5.10
pyreadline==2.1
python-dateutil==2.6.1
pytz==2017.3
pywin32==222
PyYAML==3.12
regex==2017.4.5
requests==2.14.2
requests-oauthlib==0.8.0
ruamel.yaml==0.15.35
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
simplejson==3.11.1
sip==4.19.7
six==1.10.0
sklearn==0.0
sleekxmpp==1.3.3
spacy==1.8.2
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
termcolor==1.1.0
thinc==6.5.2
toolz==0.9.0
tqdm==4.11.2
tweepy==3.5.0
Twisted==17.9.0
typing==3.5.3.0
tzlocal==1.5.1
ujson==1.35
wcwidth==0.1.7
webencodings==0.5.1
webrtcvad==2.0.10
Werkzeug==0.14.1
wikipedia==1.4.0
wrapt==1.10.11
wxpy==0.3.9.8
yarg==0.1.9
zope.interface==4.4.3
